Renamed the helper file to 'helper.js' and fixed all the references to it.
